@@ -1,3 +1,10 @@
 module t
 
-go 1.14
+go 1.20
+
+require (
+	github.com/google/gopacket v1.1.19
+	golang.org/x/net v0.0.0-20190620200207-3b0461eec859
+)
+
+require golang.org/x/sys v0.0.0-20190412213103-97732733099d // indirect
