@@ -10,7 +10,7 @@ type Api struct {
 	Topic string
 }
 
-func addMonitorAPI(api *Api) {
+func goroutineAPI(api *Api) {
 	go func() {
 		time.Sleep(1 * time.Second)
 		fmt.Println("trigger", api.Topic)
@@ -31,7 +31,7 @@ func TestClosureGoroutine(t *testing.T) {
 		// 方法2： api2 存的指针, 指向slice元素; 没有闭包bug(循环结束后，指向slice元素)
 		// api2 = &apis[i]
 		fmt.Printf("api addr:%p\n", api2)
-		addMonitorAPI(api2)
+		goroutineAPI(api2)
 	}
 	time.Sleep(2 * time.Second)
 }
@@ -43,14 +43,6 @@ VariableLoop
 	3 3 3
 
 ValueLoop
-
-	0 1 2
-
-VariableRange
-
-	2 2 2
-
-ValueRange
 
 	0 1 2
 */
@@ -84,37 +76,6 @@ func TestClosureFor(t *testing.T) {
 		}
 	}
 
-	VariableRange := func() {
-		f := make([]func(), 3)
-		for i := range f {
-			// closure over variable i
-			f[i] = func() {
-				fmt.Println(i)
-			}
-		}
-		fmt.Println("VariableRange")
-		for _, f := range f {
-			f()
-		}
-	}
-
-	ValueRange := func() {
-		f := make([]func(), 3)
-		for i := range f {
-			i := i
-			// closure over value of i
-			f[i] = func() {
-				fmt.Println(i)
-			}
-		}
-		fmt.Println("ValueRange")
-		for _, f := range f {
-			f()
-		}
-	}
-
 	VariableLoop()
 	ValueLoop()
-	VariableRange()
-	ValueRange()
 }
