@@ -1,9 +1,7 @@
 package tt
 
 import (
-	"context"
 	"database/sql"
-	"fmt"
 	"log"
 	"os"
 	"time"
@@ -50,18 +48,10 @@ func GetDb() *gorm.DB {
 	// SetConnMaxLifetime 设置了连接可复用的最大时间。
 	SqlDb.SetConnMaxLifetime(time.Hour)
 
-	beforeCreate := func(scope *gorm.DB) {
-		fmt.Println("before create sql")
-		contextScopeKey := "my_key"
-		rctx, _ := scope.Get(contextScopeKey)
-		ctx, ok := rctx.(context.Context)
-		if !ok || ctx == nil {
-			ctx = context.Background()
-		}
-		// handle ctx ....
-		scope.Set(contextScopeKey, ctx)
+	//Debug
+	if os.Getenv("DEBUG") == "true" {
+		Db = Db.Debug()
 	}
-	Db.Callback().Create().Before("gorm:create").Register("my_plugin:before_create", beforeCreate)
 
 	return Db
 }
